moved minimization methods to separate folder op_methods. moved out normalizatoin method in class Optimizer to Minimizer class
