+++ op_methods/__init__.py
@@ -1,0 +1,15 @@
+from __future__ import absolute_import, print_function
+
+__all__ = ['Simplex', 'SimplexNorm',
+           'Powell', 'CustomMinimizer',
+           'GaussProcessSKLearn', 'GaussProcess',
+           'ESMin',
+           "GPgpy"]
+
+from op_methods.simplex import *
+from op_methods.powell import *
+from op_methods.custom_minimizer import *
+from op_methods.gp_sklearn import *
+from op_methods.gp_slac import *
+from op_methods.es import *
+from op_methods.gp_gpy import *
